// mobile/lib/types.ts
export interface Address {
  id: string;
  name: string;
  phone: string;
  street: string;
  city: string;
  county: string | null;
  postalCode: string | null;
  country?: string;
  isDefault: boolean;
}